\name{exams2nops}
\alias{exams2arsnova}

\title{Generation of Exams in ARSnova Format}

\description{
  Interfaces for generating exercise sessions in the
  audience response system ARSnova.
}

\usage{
  exams2arsnova(file, n = 1L, dir = ".", name = "VU",
    quiet = TRUE, resolution = 100, width = 4, height = 4, encoding = "", converter = "ttm",
    url = "https://arsnova.uibk.ac.at/", sessionkey = NULL, jsessionid = NULL,
    active = TRUE, showstatistic = FALSE, showanswer = FALSE,
    abstention = TRUE, variant = "lecture", ssl.verifypeer = TRUE, \dots)
}

\arguments{
  \item{file}{character. A specification of a (list of) exercise files.}
  \item{n}{integer. The number of copies to be compiled from \code{file}.}
  \item{dir}{character. The default is either display on the screen or
    the current working directory.}
  \item{name}{character. A name prefix for resulting exercises and RDS file.}
  \item{quiet}{logical. Should output be suppressed when calling
    \code{\link[utils]{Sweave}} and \code{\link[tools]{texi2dvi}}.}
  \item{resolution, width, height}{numeric, passed to \code{\link[utils]{Sweave}}.}
  \item{encoding}{character, passed to \code{\link[utils]{Sweave}}.}
  \item{converter}{character passted to \code{\link[exams]{make_exercise_transform_html}}.}
  \item{url}{character specifying the base URL of the ARSnova server.}
  \item{sessionkey}{character specifying the 8-digit ARSnova session key.}
  \item{jsessionid}{character specifying the JSESSIONID cookie of an active
    ARSnova session.}
  \item{active}{logical. Should the question be active (i.e., released for students)
    or locked?}
  \item{showstatistic}{logical. Should statistics be shown?}
  \item{showanswer}{logical. Should answers be shown?}
  \item{abstention}{logical. Are abstentions?}
  \item{variant}{character. Should the question be a \code{"lecture"} or
    a \code{"preparation"} questions?}
  \item{ssl.verifypeer}{logical. Should SSL certificates be validated when
    connecting via https?}
  \item{\dots}{arguments passed on to \code{\link[exams]{xexams}}.}
}

\details{
  \code{exams2arsnova} produces JSON for ARSnova.
}

\value{
  A list of exams as generated by \code{\link[exams]{xexams}} is
  returned invisibly.
}

\examples{
## load package and enforce par(ask = FALSE)
library("exams")
options(device.ask.default = FALSE)

\dontrun{
## exams2arsnova can either create text files with JSON data
exams2arsnova("tstat2")
## or directly post this to an active ARSnova session (for which the
## server URL, the 8-digit session key, and the JSESSIONID cookie are needed)
exams2arsnova("tstat2", url = "https://arsnova.uibk.ac.at",
  sessionkey = "49061284", jsessionid = "A5BEFDA4141816BB425F2204A602E4B3")
}
}

\keyword{utilities}
