\name{exams2canvas}
\alias{exams2canvas}

\title{Generation of Exams for Canvas}

\description{
  Automatic generation of exams in QTI 1.2 with some
  tweaks optimized for Canvas.
}

\usage{
  exams2canvas(file, n = 1L, dir = ".", name = "canvasquiz",
  maxattempts = 1, converter = NULL, \dots)
}

\arguments{
  \item{file}{character. A specification of a (list of) exercise files.}
  \item{n}{integer. The number of copies to be compiled from \code{file}.}
  \item{dir}{character. The default is the current working directory.}
  \item{name}{character. A name prefix for resulting exercises and ZIP file.}
  \item{maxattempts}{integer. The maximum attempts for one question
    (must be smaller than \code{100000}).}
  \item{converter}{character passed on to \code{\link[exams]{make_exercise_transform_html}},
    indicating the type of converter from LaTeX/Markdown to HTML. Defaults
    to HTML with MathML.}
  \item{\dots}{arguments passed on to \code{\link[exams]{exams2qti12}}.}
}

\details{
  \code{exams2canvas} is a convenience interface to \code{\link[exams]{exams2qti12}}
  for generating QTI 1.2 with some small tweaks for Canvas.
}

\value{
  \code{exams2canvas} returns a list of exams as generated by \code{\link[exams]{xexams}}.
}

\seealso{
  \code{\link[exams]{exams2qti12}}
}

\examples{
\donttest{
## load package and enforce par(ask = FALSE)
library("exams")
options(device.ask.default = FALSE)

## define an exams (= list of exercises)
myexam <- c(
  "boxplots",
  "tstat",
  "ttest",
  "regression",
  "relfreq"
)

## output directory
mydir <- tempdir()

## generate .zip with OpenOLAT test in temporary directory
exams2canvas(myexam, n = 3, dir = mydir)
dir(mydir)
}
}

\keyword{utilities}
