\name{exams2evaexam}
\alias{exams2evaexam}

\title{Generation of Exam Exercises in EvaExam Format}

\description{
  Interface for generating text files for import
  in the exam system EvaExam.
}

\usage{
exams2evaexam(file, n = 1L, dir = ".", name = "evaexam",
  quiet = TRUE, resolution = 100, width = 4, height = 4,
  points = NULL, fix_choice = TRUE, sep = ";",
  converter = "pandoc-mathjax", base64 = TRUE, \dots)
}

\arguments{
  \item{file}{character. A specification of a (list of) exercise files.}
  \item{n}{integer. The number of copies to be compiled from \code{file}.}
  \item{dir}{character. The default is either display on the screen or
    the current working directory.}
  \item{name}{character. A name prefix for resulting file.}
  \item{quiet}{logical. Should output be suppressed when calling
    \code{\link[exams]{xweave}} and \code{\link[tools]{texi2dvi}}.}
  \item{resolution, width, height, points}{numeric, passed to \code{\link[exams]{xweave}}.}
  \item{fix_choice}{logical. Should math markup be removed in single and multiple
    choice lists?}
  \item{sep}{character. Separator for columns passed on to \code{\link{write.table}}.}
  \item{converter, base64, \dots}{arguments passed on to \code{make_exercise_transform_html},
    controlling the HTML formatting of the question texts.}
}

\details{
  \code{exams2evaexam} generates exams in semicolon- or comma-separated values (CSV)
  format that can be imported into the exam system EvaExam
  (\url{https://evasys.de/evaexam/}) using \code{\link[exams]{xexams}}.
  In particular, single-choice and multiple-choice exercises are fully
  supported while num and string question are converted to open-ended text
  questions.

  CAUTION: Various formatting features do not work properly yet, in particular
  images will be dropped upon import into EvaExam, and mathematical equations are
  not rendered correctly.

  The function proceeds by (1) calling \code{\link[exams]{xweave}}
  on each exercise, (2) reading the resulting Markdown/LaTeX text, (3) transforming
  the text to HTML, and (4) embedding the HTML text into the
  CSV format for EvaExam.
  For steps (1) and (2) the standard drivers in \code{xexams} are used.  
  For step (3) a suitable transformation function is set up on the fly
  using \code{\link[exams]{make_exercise_transform_pandoc}}.
  For step (4) a custom writer function is set up on the fly.
}

\value{
  A list of exams as generated by \code{\link[exams]{xexams}} is
  returned invisibly.
}

\examples{
## load package and enforce par(ask = FALSE)
library("exams")
options(device.ask.default = FALSE)

## output directory
dir.create(mydir <- tempfile())

## create a CSV file evaexam-1.csv for import in EvaExam
exams2evaexam(c("swisscapital.Rmd", "capitals.Rmd", "deriv2.Rmd",
  "boxplots.Rmd", "ttest.Rmd", "function.Rmd"), dir = mydir)
}

\keyword{utilities}
