\name{exams2inspera}
\alias{exams2inspera}

\title{Generation of Exams for Inspera}

\description{
  Automatic generation of exams in QTI 2.1 with some tweaks optimized for Inspera.
}

\usage{
  exams2inspera(file, n = 1L, dir = ".", name = "inspera", template = "inspera",
    converter = "pandoc-mathml", base64 = FALSE, \dots)
}

\arguments{
  \item{file}{character. A specification of a (list of) exercise files.}
  \item{n}{integer. The number of copies to be compiled from \code{file}.}
  \item{dir}{character. The default is the current working directory.}
  \item{name}{character. A name prefix for resulting exercises and ZIP file.}
  \item{template}{character. The QTI 2.1 template that should be used, defaulting
    to \code{"inspera.xml"} provided within the package.}
  \item{converter}{character passed on to \code{\link[exams]{make_exercise_transform_html}},
    indicating the type of converter from LaTeX/Markdown to HTML. Defaults
    to HTML with MathML.}
  \item{base64}{logical. Should supplementary files be embedded using Base 64 coding?
    Currently only \code{FALSE} is supported by Inspera.}
  \item{\dots}{arguments passed on to \code{\link[exams]{exams2qti21}}, providing
    many further arguments for customization.}
}

\details{
  \code{exams2inspera} is a convenience interface to \code{\link[exams]{exams2qti21}}
  for generating QTI 2.1 output with some small tweaks for Inspera. Most importantly,
  no full "assessment" is exported but only a collection of individual exercises.
  Consequently, all options from \code{exams2qti21} that pertain to the configuration
  of the assessment (e.g., \code{maxattempts}, \code{cutvalue}, \code{navigation}, etc.)
  have no effect here. Moreover, MathJax is used for type-setting mathematical content
  by default.
}

\value{
  \code{exams2inspera} returns a list of exams as generated by \code{\link[exams]{xexams}}.
}

\seealso{
  \code{\link[exams]{exams2qti21}}
}

\examples{
## load package and enforce par(ask = FALSE)
library("exams")
options(device.ask.default = FALSE)

## define an exams (= list of exercises)
myexam <- list(
  "boxplots.Rmd",
  c("tstat.Rmd", "ttest.Rmd", "confint2.Rmd"),
  c("regression.Rmd", "anova.Rmd"),
  c("scatterplot.Rmd", "boxhist.Rmd"),
  "relfreq.Rmd"
)

## output directory
dir.create(mydir <- tempfile())

## generate .zip with exercises for Inspera in temporary directory
exams2inspera(myexam, n = 3, dir = mydir)
dir(mydir)
}

\keyword{utilities}
