\name{exams2nops}
\alias{exams2nops}
\alias{make_nops_template}

\title{Generation of Written Exams for Automatic Evaluation}

\description{
  Generation of exams in PDF format that can be printed, scanned,
  and evaluated automatically.
}

\usage{
  exams2nops(file, n = 1L, dir = NULL, name = NULL,
    language = "en", title = "Exam", course = "",
    institution = "R University", logo = "Rlogo.png", date = Sys.Date(), 
    replacement = FALSE, intro = NULL, blank = NULL, duplex = TRUE, pages = NULL,
    usepackage = NULL, encoding = "", startid = 1L, showpoints = FALSE, \dots)

  make_nops_template(n, replacement = FALSE, intro = NULL, blank = NULL,
    duplex = TRUE, pages = NULL, file = NULL, nchoice = 5, encoding = "")
}

\arguments{
  \item{file}{character. A specification of a (list of) exercise files.}
  \item{n}{integer. The number of copies to be compiled from \code{file}
    (in \code{exams2nops}) and the number of exercises per exam
    (in \code{make_nops_template}), respectively.}
  \item{dir}{character. The default is either display on the screen or
    the current working directory.}
  \item{name}{character. A name prefix for resulting exams and RDS file.}
  \item{language}{character. Path to a DCF file with a language specification.
    Currently, \code{"en"} and \code{"de"} are shipped with the package.}
  \item{title}{character. Title of the exam, e.g., \code{"Introduction to Statistics"}.}
  \item{course}{character. Optional course number, e.g., \code{"101"}.}
  \item{institution}{character. Name of the institution at which the exam is conducted.}
  \item{logo}{character. Path to a logo image. If the logo is not found, it is
    simply omitted.}
  \item{date}{character or \code{"Date"} object specifying the date of the exam.}
  \item{replacement}{logical. Should a replacement exam sheet be included?}
  \item{intro}{character with LaTeX code for optional introduction text on the first page
    of the exam.}
  \item{blank}{integer. Number of blank pages to be added at the end.
    (Default is chosen to be half of the number of exercises.)}
  \item{duplex}{logical. Should blank pages be added after the title page
    (for duplex printing)?}
  \item{pages}{character. Path(s) to additional PDF pages to be included
    at the end of the exam (e.g., formulary or distribution tables).}
  \item{usepackage}{character. Names of additional LaTeX packages to be
    included.}
  \item{encoding}{character, passed to \code{\link[exams]{xweave}}.}
  \item{startid}{integer. Starting ID for the exam numbers (defaults to 1).}
  \item{showpoints}{logical. Should the number of points associated with
    each exercise (if specified) be shown in the PDF?}
  \item{\dots}{arguments passed on to \code{\link[exams]{exams2pdf}}.}
  \item{nchoice}{character. The number of choice alternatives per exercise.}
}

\details{
  \code{exams2nops} is a convenience interface for \code{\link[exams]{exams2pdf}}
  with a dynamically generated title page which can be scanned and evaluated
  automatically.
}

\value{
  A list of exams as generated by \code{\link[exams]{xexams}} is
  returned invisibly.
}

\examples{
## load package and enforce par(ask = FALSE)
library("exams")
options(device.ask.default = FALSE)

## define an exams (= list of exercises)
myexam <- list(
  "boxplots",
  "ttest",
  "anova",
  "scatterplot",
  "relfreq",
  "tstat2"
)

## output directory
mydir <- tempdir()

## generate NOPS exam in temporary directory
## using a few customization options
exams2nops(myexam, n = 3, dir = mydir, replacement = TRUE)
dir(mydir)

if(interactive()) {
## compile a single random exam (displayed on screen)
exams2nops(myexam, duplex = FALSE, language = "de")
}
}

\keyword{utilities}
