\name{exams2wu}
\alias{exams2wu}
\alias{make_exams_write_wu}

\title{Generation of Exams in WU Wien Exam Server Format}

\description{
  Automatic generation of exams in WU Wien exam server format.
}

\usage{
  exams2wu(file, n = 1L, nsamp = NULL, dir = NULL, name = NULL, 
    quiet = TRUE, edir = NULL, tdir = NULL, sdir = NULL, solution = TRUE, 
    doctype = NULL, head = NULL, resolution = 100, width = 4, 
    height = 4, converter = "tex2image", base64 = FALSE, \dots)

  make_exams_write_wu(name = NULL, \dots)
}

\arguments{
  \item{file}{character. A specification of a (list of) exercise files.}
  \item{n}{integer. The number of copies to be compiled from \code{file}.}
  \item{nsamp}{integer. The number of exercise files sampled from each
    list element of \code{file}. Sampling without replacement is used
    if possible. (Only if some element of \code{nsamp} is larger than
    the length of the corresponding element in \code{file}, sampling
    with replacement is used.)}
  \item{dir}{character. The output directory, this has to be set if
    \code{n} is greater than 1.}
  \item{name}{character. A name prefix for resulting exercises.}
  \item{quiet}{logical. Should output be suppressed when calling
    \code{\link[utils]{Sweave}}?}
  \item{edir}{character specifying the path of the directory in which
    the files in \code{file} are stored (see also below).}
  \item{tdir}{character specifying a temporary directory, by default
    this is chosen via \code{\link[base]{tempdir}}.}
  \item{sdir}{character specifying a directory for storing supplements, by
    default this is chosen via \code{\link[base]{tempdir}}.}
  \item{solution}{logical. Should the solution be included in the HTML
    output?}
  \item{doctype}{character vector with a DOCTYPE tag for the HTML page.
    By default HTML4 is employed.}
  \item{head}{character vector for the head tag. By default a simple
    header is employed, setting the font to Arial.}
  \item{resolution, width, height}{numeric. Options for rendering PNG
    graphics passed to \code{\link[utils]{Sweave}}.}
  \item{converter}{character. Workhorse function for transforming LaTeX
    code to HTML.}
  \item{base64}{logical. Should images be embeddeg using Base 64 coding?}    
  \item{\dots}{arguments passed on to \code{make_exercise_transform_html}.}
}

\details{
  Still to do.
}

\value{
  \code{exams2wu} returns a list of exams as generated by \code{\link[exams]{xexams}}.
  
  \code{make_exams_write_wu} ... still to do.
}

\seealso{
  \code{\link[exams]{xexams}},
  \code{\link[tth]{ttm}},
  \code{\link[tth]{tth}},
  \code{\link[exams]{tex2image}},
  \code{\link[exams]{make_exercise_transform_html}},
}

\examples{
\dontrun{
## still to do
}
}

\keyword{utilities}
