% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyExams.R
\name{runShinyExams}
\alias{runShinyExams}
\title{Run shinyExams}
\usage{
runShinyExams(...)
}
\arguments{
\item{...}{[\code{any}]\cr
Additional arguments passed to shiny's
\code{runApp()} function.}
}
\description{
Run a local instance of shinyExams.
}
\examples{
\dontrun{
  runShinyRexams()
}
}
