\name{tex2image}
\alias{tex2image}

\title{Transforming LaTeX Code Using ImageMagick}

\description{
  Transformation of LaTeX code into an image by compiling
  to PDF and then transforming to PNG via ImageMagick.
}

\usage{
  tex2image(tex, format = "png", width = 6, pt = 12, density = 350, 
    edir = NULL, tdir = NULL, idir = NULL, width.border = 0L, 
    col.border = "white", resize = 650, shave = 4, template, 
    itemplate = NULL, show = TRUE, bsname = "tex2image-Rinternal", 
    \dots)
}

\arguments{
  Still to do.
}

\details{
  Still to do.
}

\value{
  Still to do.
}

\seealso{
  \code{\link[tools]{texi2dvi}}
}

\examples{
## still to do
}

\keyword{utilities}
