\name{exams2canvas}
\alias{exams2canvas}

\title{Generation of Exams for Canvas}

\description{
  Automatic generation of exams in QTI 1.2 with some
  tweaks optimized for Canvas.
}

\usage{
  exams2canvas(file, n = 1L, dir = ".", name = "canvasquiz",
    maxattempts = 1, duration = NULL, points = NULL,
    converter = NULL, \dots)
}

\arguments{
  \item{file}{character. A specification of a (list of) exercise files.}
  \item{n}{integer. The number of copies to be compiled from \code{file}.}
  \item{dir}{character. The default is the current working directory.}
  \item{name}{character. A name prefix for resulting exercises and ZIP file.}
  \item{maxattempts}{integer. The maximum attempts for one question
    (must be smaller than \code{100000}).}
  \item{duration}{integer. Set the duration of the exam in minutes.}
  \item{points}{integer. How many points should be assigned to each exercise? Note that this
    argument overules any exercise points that are provided within an \code{"\expoints{}"} tag
    in the \code{.Rnw} file. The vector of points supplied is expanded to the number of exercises in
    the exam.}
  \item{converter}{character passed on to \code{\link[exams]{make_exercise_transform_html}},
    indicating the type of converter from LaTeX/Markdown to HTML. Currently,
    Canvas requires HTML with MathML. Hence, if there are Markdown-based .Rmd
    exercises, the \code{converter} needs to be \code{"pandoc-mathml"} (or
    equivalently \code{"pandoc"}). For only LaTeX-based .Rnw exercises \code{"ttm"}
    can be used as an alternative and is currently the default.}
  \item{\dots}{arguments passed on to \code{\link[exams]{exams2qti12}}.}
}

\details{
  \code{exams2canvas} is a convenience interface to \code{\link[exams]{exams2qti12}}
  for generating QTI 1.2 with some small tweaks for Canvas. In particular, Canvas
  necessitates a \code{converter} that produces HTML plus MathML and supplementary
  files (images, data, ...) must be embedded without Base 64. Finally, the QTI XML
  file requires a few special tags which are enforced through \code{flavor = "canvas"}
  in \code{exams2qti12}.

  Note that Canvas does not currently allow partial credits on mchoice questions, according
  to the documentation, but an evaluation rule with partial credits is still applied. Therefore
  there is a warning if the evaluation rule is all or nothing (i.e., \code{partial = FALSE}).
  Moreover, negative points are currently no supprted by Canvas.
}

\value{
  \code{exams2canvas} returns a list of exams as generated by \code{\link[exams]{xexams}}.
}

\seealso{
  \code{\link[exams]{exams2qti12}}
}

\examples{
\donttest{
## load package and enforce par(ask = FALSE)
library("exams")
options(device.ask.default = FALSE)

## define an exams (= list of exercises)
myexam <- c(
  "boxplots",
  "tstat",
  "ttest",
  "regression",
  "relfreq"
)

## output directory
mydir <- tempdir()

## generate .zip with OpenOLAT test in temporary directory
exams2canvas(myexam, n = 3, dir = mydir)
dir(mydir)
}
}

\keyword{utilities}
