\name{exams2imsqti}
\alias{exams2imsqti}
\alias{make_exams_write_imsqti}

\title{Generation of Exams in IMS QTI 1.2 Format}

\description{
  Automatic generation of exams in IMS QTI 1.2 format.
}

\usage{
  exams2imsqti(file, n = 1L, nsamp = NULL, dir = NULL, name = NULL, 
    quiet = TRUE, edir = NULL, tdir = NULL, sdir = NULL, resolution = 100, 
    width = 4, height = 4, \dots)

  make_exams_write_imsqti(x, dir, tdir = NULL, name = NULL,
    template.assessment = NULL, template.item = NULL, \dots)
}

\arguments{
  \item{file}{character. A specification of a (list of) exercise files.}
  \item{n}{integer. The number of copies to be compiled from \code{file}.}
  \item{nsamp}{integer. The number of exercise files sampled from each
    list element of \code{file}. Sampling without replacement is used
    if possible. (Only if some element of \code{nsamp} is larger than
    the length of the corresponding element in \code{file}, sampling
    with replacement is used.)}
  \item{dir}{character. The output directory, this has to be set if
    \code{n} is greater than 1.}
  \item{name}{character. A name prefix for resulting exercises.}
  \item{quiet}{logical. Should output be suppressed when calling
    \code{\link[utils]{Sweave}}?}
  \item{edir}{character specifying the path of the directory in which
    the files in \code{file} are stored (see also below).}
  \item{tdir}{character specifying a temporary directory, by default
    this is chosen via \code{\link[base]{tempdir}}.}
  \item{sdir}{character specifying a directory for storing supplements, by
    default this is chosen via \code{\link[base]{tempdir}}.}
  \item{resolution, width, height}{numeric. Options for rendering PNG
    graphics passed to \code{\link[utils]{Sweave}}.}

  \item{x}{ still to do. }  
  \item{template.assessment}{ still to do. }
  \item{template.item}{ still to do. }
    
  \item{\dots}{arguments passed on to \code{make_exercise_transform_html}.}
}

\details{
  Still to do.
}

\value{
  \code{exams2imsqti} returns a list of exams as generated by \code{\link[exams]{xexams}}.
  
  \code{make_exams_write_imsqti} ... still to do.
}

\seealso{
  \code{\link[exams]{xexams}},
  \code{\link[tth]{ttm}},
  \code{\link[tth]{tth}},
  \code{\link[exams]{tex2image}},
  \code{\link[exams]{make_exercise_transform_html}},
}

\examples{
## still to do
}

\keyword{utilities}
