\name{exams2qti12}
\alias{exams2qti12}
\alias{make_itembody_cloze_qti12}
\alias{make_itembody_num_qti12}
\alias{make_itembody_num_olat}
\alias{make_itembody_mchoice_qti12}
\alias{make_itembody_schoice_qti12}

\title{Generation of Exams in IMS QTI 1.2 Format}

\description{
  Automatic generation of exams in IMS QTI 1.2 format.
}

\usage{
  exams2qti12(file, n = 1L, nsamp = NULL, dir,
    name = NULL, quiet = TRUE, edir = NULL,
    tdir = NULL, sdir = NULL, resolution = 100,
    width = 4, height = 4, num = NULL, mchoice = NULL,
    schoice = mchoice, cloze = NULL, template = NULL, \dots)

  make_itembody_cloze_qti12(defaultval = NULL, minvalue = NULL,
    maxvalue = NULL, cutvalue = NULL, lang = "en", digits = 2,
    enumerate = TRUE)

  make_itembody_num_qti12(defaultval = NULL, minvalue = NULL,
    maxvalue = NULL, cutvalue = NULL)

  make_itembody_num_olat(defaultval = NULL, minvalue = NULL,
    maxvalue = NULL, cutvalue = NULL, digits = 2)

  make_itembody_mchoice_qti12(rtiming = NULL, shuffle = FALSE,
    minnumber = NULL, maxnumber = NULL, rshuffle = shuffle,
    defaultval = NULL, minvalue = NULL, maxvalue = NULL,
    cutvalue = NULL, enumerate = TRUE)

  make_itembody_schoice_qti12(rtiming = NULL, shuffle = FALSE,
    minnumber = NULL, maxnumber = NULL, rshuffle = shuffle,
    defaultval = NULL, minvalue = NULL, maxvalue = NULL,
    cutvalue = NULL, enumerate = TRUE)
}

\arguments{
  \item{file}{character. A specification of a (list of) exercise files.}
  \item{n}{integer. The number of copies to be compiled from \code{file}.}
  \item{nsamp}{integer. The number of exercise files sampled from each
    list element of \code{file}. Sampling without replacement is used
    if possible. (Only if some element of \code{nsamp} is larger than
    the length of the corresponding element in \code{file}, sampling
    with replacement is used.)}
  \item{dir}{character. The output directory, this has to be set if
    \code{n} is greater than 1.}
  \item{name}{character. A name prefix for resulting exercises.}
  \item{quiet}{logical. Should output be suppressed when calling
    \code{\link[utils]{Sweave}}?}
  \item{edir}{character specifying the path of the directory in which
    the files in \code{file} are stored (see also below).}
  \item{tdir}{character specifying a temporary directory, by default
    this is chosen via \code{\link[base]{tempdir}}.}
  \item{sdir}{character specifying a directory for storing supplements, by
    default this is chosen via \code{\link[base]{tempdir}}.}
  \item{resolution, width, height}{numeric. Options for rendering PNG
    graphics passed to \code{\link[utils]{Sweave}}.}
  \item{num}{function or named list applied to numerical (i.e., type \code{num})
    questions. If \code{num} is a function, \code{num} will be used
    for generating the item body of the question, see function \code{make_itembody_num()}.
    If \code{num} is a named list, these arguments will be passed to function
    \code{make_itembody_num()}.}
  \item{mchoice, schoice, cloze}{function or named list applied to multiple choice, 
    single choice and cloze questions (i.e., type \code{mchoice}, \code{schoice}, \code{cloze}),
    respectively. See argument \code{num} for more details.}
  \item{template}{character. The IMS QTI template that should be used.}
  \item{enumerate, lang}{item specific stuff, still to do.}
  \item{rtiming, shuffle, minnumber, maxnumber, rshuffle, defaultval, minvalue,
    maxvalue, cutvalue}{arguments used for IMS QTI 1.2 item construction, for details see the XML 
    specification (see IMS Global Learning Consortium, Inc. 2012), especially Section 4.}
  \item{digits}{integer. how many digits should be used?}
  \item{\dots}{arguments passed on to \code{make_exercise_transform_html}.}
}

\details{
  Still to do.
}

\references{
  IMS Global Learning Consortium, Inc. (2012).
  \emph{IMS Question & Test Interoperability: ASI XML Binding Specification Final Specification Version 1.2}.
  \url{http://www.imsglobal.org/question/qtiv1p2/imsqti_asi_bindv1p2.html}
}

\value{
  \code{exams2qti12} returns a list of exams as generated by \code{\link[exams]{xexams}}.
  
  \code{make_exams_write_qti12} ... still to do.
}

\seealso{
  \code{\link[exams]{xexams}},
  \code{\link[tth]{ttm}},
  \code{\link[tth]{tth}},
  \code{\link[exams]{tex2image}},
  \code{\link[exams]{make_exercise_transform_html}},
}

\examples{
## still to do
}

\keyword{utilities}
