\name{tex2image}
\alias{tex2image}

\title{Transforming LaTeX Code Using ImageMagick}

\description{
  Transformation of LaTeX code into an image by compiling
  to PDF and then transforming to PNG via ImageMagick.
}

\usage{
  tex2image(tex, format = "png", width = 6, pt = 12, density = 350, 
    edir = NULL, tdir = NULL, idir = NULL, width.border = 0L, 
    col.border = "white", resize = 650, shave = 4, template, 
    itemplate = NULL, show = TRUE, bsname = "tex2image-Rinternal", 
    \dots)
}

\arguments{
  \item{tex}{ still to do. }
  \item{format}{ still to do. }
  \item{width}{ still to do. }
  \item{pt}{ still to do. }
  \item{density}{ still to do. }
  \item{edir}{ still to do. }
  \item{tdir}{ still to do. }
  \item{idir}{ still to do. }
  \item{width.border}{ still to do. }
  \item{col.border}{ still to do. }
  \item{resize}{ still to do. }
  \item{shave}{ still to do. }
  \item{template}{ still to do. }
  \item{itemplate}{ still to do. }
  \item{show}{ still to do. }
  \item{bsname}{ still to do. }
  \item{\dots}{ still to do. }
}

\details{
  Still to do.
}

\value{
  Still to do.
}

\seealso{
  \code{\link[tools]{texi2dvi}}
}

\examples{
## still to do
}

\keyword{utilities}
